requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
chardet==5.2.0
crispy-bootstrap5==2024.10
Django==4.2.18
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
django-widget-tweaks==1.5.0
django_widgets==0.1.16
dnspython==2.7.0
et_xmlfile==2.0.0
mysqlclient==2.2.7
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
phonenumbers==9.0.5
pillow==11.1.0
py3dns==4.0.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
reportlab==4.4.1
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tzdata==2024.2
Unidecode==1.4.0
validate-email-address==1
whitenoise==6.6.0
xlsxwriter==3.2.5
